Implemented "binding" on password field Now possible to save and add new users, but must type password every time. Still TODO: proper exception handling, right now the whole app will outright crash due to usage of Commands

+++ Arvato.TestProject.UsrMgmt.UI.Desktop/ViewModels/UsersListViewModel.cs
@@ -81,7 +81,7 @@ namespace Arvato.TestProject.UsrMgmt.UI.Desktop.ViewModels
 
         private void RefreshUsers()
         {
-            users = userService.GetList();
+            Users = userService.GetList();
         }
 
         #region Command methods
@@ -93,6 +93,17 @@ namespace Arvato.TestProject.UsrMgmt.UI.Desktop.ViewModels
 
         private void DeleteUser()
         {
+            var result = MessageBox.Show(
+                String.Format(@"Are you sure you want to delete user ""{0}""?", FormViewModel.CurrentUser.LoginID),
+                "Deleting user",
+                MessageBoxButton.YesNo,
+                MessageBoxImage.Warning);
+
+            if (result == MessageBoxResult.No)
+            {
+                return;
+            }
+
             try
             {
                 userService.Delete(FormViewModel.CurrentUser);
